app/requirements.txt
Flask==1.1.1
psycopg2-binary
pandas
xlrd >= 1.0.0
